Move rust bindings out of suricata-common, allow gen to be run at anytime
