+++ src/app-layer-dhcp.c
@@ -25,6 +25,7 @@
 #include "util-unittest.h"
 #include "app-layer-parser.h"
 #include "app-layer-dhcp.h"
+#include "rust.h"
 
 void RegisterDHCPParsers(void)
 {
