+++ src/app-layer-ftp.c
@@ -60,6 +60,7 @@
 #include "util-misc.h"
 
 #include "output-json.h"
+#include "rust.h"
 
 typedef struct FTPThreadCtx_ {
     MpmThreadCtx *ftp_mpm_thread_ctx;
