+++ src/app-layer-sip.c
@@ -33,6 +33,7 @@
 #include "app-layer-parser.h"
 
 #include "app-layer-sip.h"
+#include "rust.h"
 
 void RegisterSIPParsers(void)
 {
