+++ src/detect-snmp-pdu_type.c
@@ -29,6 +29,7 @@
 #include "detect-engine-content-inspection.h"
 #include "detect-snmp-pdu_type.h"
 #include "app-layer-parser.h"
+#include "rust.h"
 
 /**
  *   [snmp.pdu_type]:<type>;
