+++ src/detect-snmp-version.c
@@ -29,6 +29,7 @@
 #include "detect-engine-content-inspection.h"
 #include "detect-snmp-version.h"
 #include "app-layer-parser.h"
+#include "rust.h"
 
 /**
  *   [snmp.version]:[<|>|<=|>=]<version>;
