+++ src/detect-template-rust-buffer.c
@@ -38,6 +38,7 @@
 #include "detect-engine-content-inspection.h"
 #include "detect-template-rust-buffer.h"
 #include "app-layer-parser.h"
+#include "rust.h"
 
 
 static int DetectTemplateRustBufferSetup(DetectEngineCtx *, Signature *,
