+++ src/output-json-dhcp.c
@@ -44,6 +44,7 @@
 
 #include "app-layer-dhcp.h"
 #include "output-json-dhcp.h"
+#include "rust.h"
 
 
 typedef struct LogDHCPFileCtx_ {
