+++ src/output-json-dns.c
@@ -50,6 +50,7 @@
 
 #include "output-json.h"
 #include "output-json-dns.h"
+#include "rust.h"
 
 /* we can do query logging as well, but it's disabled for now as the
  * TX id handling doesn't expect it */
