+++ src/output-json-rdp.c
@@ -41,6 +41,7 @@
 #include "app-layer-parser.h"
 #include "app-layer-rdp.h"
 #include "output-json-rdp.h"
+#include "rust.h"
 
 typedef struct LogRdpFileCtx_ {
     LogFileCtx *file_ctx;
