+++ src/output-json-template-rust.c
@@ -52,7 +52,7 @@
 
 #include "app-layer-template-rust.h"
 #include "output-json-template-rust.h"
-
+#include "rust.h"
 
 typedef struct LogTemplateFileCtx_ {
     LogFileCtx *file_ctx;
