+++ src/util-debug.c
@@ -42,6 +42,7 @@
 
 #include "util-unittest.h"
 #include "util-syslog.h"
+#include "rust.h"
 
 
 #include "conf.h"
